requirements.txt
ephem
python-telegram-bot==13.15
